Create `MockRpc{Client, Server, ServerApi}`
To replace the previous test helper types. This also changes how it
operates. Now there is no internal delay in the operations, there is an
operation that may simulate it taking too long instead.
